pyproject.toml
[tool.poetry]
name = "Bank_of_Thailand_Exchange_Rate"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "Bank_of_Thailand_Exchange_Rate"},
    { include = "tests", format = "sdist"},
]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
requests = "^2.31.0"
pandas = "^2.0.0"
pyarrow = "^15.0.0"
pytest = "6.2.5"
apache-airflow = "2.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.1"
flake8-annotations = "^2.9.1"
isort = "^5.12.0"
black = "^23.3.0"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
   \.git
  | \.tox
  | build
  | migrations
  | .venv
)/
'''

[tool.flake8]
max-line-length = 120

[tool.isort]
profile = 'black'
